ISPN-2131 Test cleanup: remove unused test entity, avoid some autoboxing, remove unused imports

+++ query/src/main/java/org/infinispan/query/CommandFactory.java
@@ -40,7 +40,7 @@ public class CommandFactory implements ExtendedModuleCommandFactory {
    @Override
    public Map<Byte, Class<? extends ReplicableCommand>> getModuleCommands() {
       Map<Byte, Class<? extends ReplicableCommand>> map = new HashMap<Byte, Class<? extends ReplicableCommand>>(1);
-      map.put(ClusteredQueryCommand.COMMAND_ID, ClusteredQueryCommand.class);
+      map.put(Byte.valueOf(ClusteredQueryCommand.COMMAND_ID), ClusteredQueryCommand.class);
       return map;
    }
